try to improve docs and mod organization
