+++ kernel/src/coroutine/mod.rs
@@ -1,30 +1,126 @@
-//! Stackless, connector-driven kernel workflows.
+//! Stackless, connector-driven kernel coroutines.
 //!
-//! By design, kernel maintains a strict separation of concerns, delegating I/O and dataflow
-//! operations to the connector. The [`crate::Engine`] traits allow kernel to delegate work, but
-//! they produce nested connector -> kernel -> connector call stacks with kernel stuck in the middle
-//! between two sides of the connector, which complicates the connector's state passing, scheduling,
-//! error handling, etc. It also creates a high risk of control inversion, where kernel code forces
-//! design decisions that rightfully belong with the connector, by treating connector as a library
-//! to be called upon at will. Stackless coroutines avoid all of that: Whenever kernel needs to
-//! communicate back to the connector, it suspends itself and the connector receives the request as
-//! a normal function call return. The connector responds to a request by running its own code on
-//! behalf of kernel and resuming kernel with the response. Kernel coroutines take two forms: A
-//! [`Workflow`] evolves through multiple request-response steps to produce a final answer, while a
-//! [`Generator`] can additionally yield items before producing its final answer. Whenever a
-//! connector starts or resumes either kind of coroutine, kernel runs on the calling thread until it
-//! either completes or suspends with the next request/yield. Because these are stackless
-//! coroutines, no kernel functions are in the call stack while kernel is suspended.
+//! When a connector invokes a kernel operation such as snapshot creation, kernel delegates all I/O
+//! and dataflow operations back to the connector. However, delegation by calling back into
+//! connector code would produce nested connector -> kernel -> connector call stacks, with connector
+//! blocking on kernel while kernel invokes connector code:
 //!
-//! The exhaustive [`Request`] enum defines the operations kernel may delegate. Each request carries
-//! the operation inputs and a [`TypedResume`] handle, which owns the suspended continuation and
-//! which declares the expected response type. A connector drives a [`Workflow`] or [`Generator`] by
-//! repeatedly matching each request, performing the operation, and invoking [`TypedResume::resume`]
-//! with the result:
+//! ```text
+//!   +---------------------------------------------------------------------------+
+//!   |                                   CONNECTOR                               |
+//!   |                                                                           |
+//!   |                        +-----------------------------+                    |
+//!   |                        |                             |                    |
+//!   |                        |     +-----KERNEL------+     |                    |
+//! ----> analyze_query        |     |                 |     |                    |
+//!   |     |                  |     |                 |     |                    |
+//!   |     +--------------------------> load_snapshot |     |                    |
+//!   |     .                  |     |        |        |     |                    |
+//!   |     .                  |     |        +----------------> list             |
+//!   |     .                  |     |        .        |     |    |               |
+//!   |     .                  |     |        .        |     |    |               |
+//!   |     .                  |     |        + <-----------------+               |
+//!   |     .                  |     |        |        |     |                    |
+//!   |     .                  |     |        +----------------> read_json        |
+//!   |     .                  |     |        .        |     |    |               |
+//!   |     .                  |     |        .        |     |    |               |
+//!   |     .                  |     |        + <-----------------+               |
+//!   |     .                  |     |        |        |     |                    |
+//!   |     .                  |     |        +----------------> read_parquet     |
+//!   |     .                  |     |        .        |     |    |               |
+//!   |     .                  |     |        .        |     |    |               |
+//!   |     .                  |     |        + <-----------------+               |
+//!   |     .                  |     |        |        |     |                    |
+//!   |     + <-----(snapshot return value)---+        |     |                    |
+//!   |     |                  |     |                 |     |                    |
+//! <-------+                  |     |                 |     |                    |
+//!   |                        |     |                 |     |                    |
+//!   +------------------------+     +-----------------+     +--------------------+
+//! ```
+//!
+//! Because kernel is wedged between two "halves" of the connector, it interferes with connector's
+//! scheduling model (sync/async/parallel/distributed), error handling, etc. It also depends on the
+//! complex [`Engine`](crate::Engine) trait hierarchy.
+//!
+//! Coroutines flip the control flow around by modeling kernel's work as a resumable continuation
+//! function. Connector still invokes a kernel function to start running kernel, but kernel
+//! delegates requests back to the connector by suspending itself. When that happens, the
+//! connector's function call returns with a (request, continuation) pair as a return value, and
+//! connector inovkes the continuation function to pass a response to kernel and allow it to
+//! continue. The process repeats until the coroutine finishes and returns the requested value:
+//!
+//! ```text
+//!   +------------------------+     +-----------------+
+//!   |       CONNECTOR        |     |     KERNEL      |
+//!   |                        |     |                 |
+//! ----> analyze_query        |     |                 |
+//!   |     |                  |     |                 |
+//!   |     +--------------------------> load_snapshot |
+//!   |     .                  |     |        |        |
+//!   |     + <----------(request listing)----S        |
+//!   |     |                  |     |        .        |
+//!   |     +---> list         |     |        .        |
+//!   |            |           |     |        .        |
+//!   |            |           |     |        .        |
+//!   |     + <----+           |     |        .        |
+//!   |     |                  |     |        .        |
+//!   |     +---(listing response)----------> R        |
+//!   |     .                  |     |        |        |
+//!   |     + <--------(request JSON read)----S        |
+//!   |     |                  |     |        .        |
+//!   |     +---> read_json    |     |        .        |
+//!   |            |           |     |        .        |
+//!   |            |           |     |        .        |
+//!   |     + <----+           |     |        .        |
+//!   |     |                  |     |        .        |
+//!   |     +---(JSON data response)--------> R        |
+//!   |     .                  |     |        |        |
+//!   |     + <------(request parquet read)---S        |
+//!   |     |                  |     |        .        |
+//!   |     +---> read_parquet |     |        .        |
+//!   |            |           |     |        .        |
+//!   |            |           |     |        .        |
+//!   |     + <----+           |     |        .        |
+//!   |     |                  |     |        .        |
+//!   |     +---(parquet data response)-----> R        |
+//!   |     .                  |     |        |        |
+//!   |     + <-----(snapshot return value)---+        |
+//!   |     |                  |     |                 |
+//! <-------+                  |     |                 |
+//!   |                        |     |                 |
+//!   +------------------------+     +-----------------+
+//! ```
+//!
+//! This way, the connector always invokes its own code, acting on behalf of kenrel as needd. The
+//! coroutines are "stackless" because there is no kernel code in the call stack while the connector
+//! processes a kernel request. When the connector starts or resumes a kernel coroutine, kernel runs
+//! synchronously in the calling thread until it either completes or suspends again (which happens
+//! quickly because kernel delegates all slow or heavy work to the connector). At each step, the
+//! connector decides how to serve the request and whether to resume kernel.
+//!
+//! Kernel coroutines come in two flavors:
+//! * A [`Workflow`] suspends zero or more times before producing a final output
+//! * A [`Generator`] is like a workflow, but can additionally yield zero or more items back to the
+//!   connector before producing its final output.
+//!
+//! Kernel coroutines are an async-friendly synchronous API: Between suspensions, kernel performs
+//! only light, bounded CPU work. It delegates all blocking, heavy, or unbounded work to the
+//! connector as one or more [`Request`] instances. Connectors therefore do not need to treat kernel
+//! calls as scheduling points, and they can schedule requests like any other work. Resume handles
+//! are [`Send`] and may move freely between tasks and threads as the workflow progresses. The
+//! connector may execute requests synchronously, asynchronously, in parallel with threads, or
+//! across distributed workers; and it can use the runtimes and libraries of its choosing to do so.
+//!
+//! The [`Request`] enum defines the operations kernel may delegate. Each request carries the
+//! operation inputs and a [`Resume`] function, which owns the suspended continuation and which
+//! declares the expected response type. A connector drives a [`Workflow`] or [`Generator`] by
+//! repeatedly matching each request, performing the operation, and resuming kernel with the result.
+//!
+//! For example, if kernel defines this:
 //!
 //! ```no_run
 //! use delta_kernel::coroutine::listing::ForwardListing;
-//! use delta_kernel::coroutine::{Cursor, Page, PageRequest, Workflow};
+//! use delta_kernel::coroutine::{PageRequest, Workflow};
 //! use delta_kernel::DeltaResult;
 //!
 //! # /*
@@ -37,7 +133,7 @@
 //!     /// List page ranges from high to low, with entries ascending within each page.
 //!     ListBackward(PageRequest<N, BackwardListing>),
 //!     /// Read a whole file when the range is `None`, or exactly the half-open range otherwise.
-//!     ReadSmallFile(FileSlice, TypedResume<N, Bytes>),
+//!     ReadSmallFile(FileSlice, Resume<N, Bytes>),
 //!       ...
 //! }
 //! # */
@@ -48,28 +144,39 @@
 //!
 //! // A public kernel workflow entry point. It does _not_ take an `Engine` instance.
 //! pub fn public_kernel_entry_point() -> DeltaResult<Workflow<Foo>> {
-//!     // Kernel constructs and starts its coroutine here.
-//!     # Err(delta_kernel::Error::generic("example-only entry point"))
+//!     // Kernel constructs and starts its coroutine here, returning to the caller
+//!     // when it needs to make a Request or has a final result ready.
+//!     todo!()
 //! }
+//! ```
+//!
+//! Then a connector would drive the `public_kernel_entry_point` workflow as follows:
+//!
+//! ```no_run
+//! use delta_kernel::coroutine::listing::ForwardListing;
+//! use delta_kernel::coroutine::{Cursor, Page, PageRequest, Request, Workflow};
+//! use delta_kernel::DeltaResult;
+//! # pub struct Foo;
+//! #
+//! # pub fn public_kernel_entry_point() -> DeltaResult<Workflow<Foo>> {
+//! #    todo!()
+//! # }
 //!
 //! async fn prepare_listing(
-//!     _request: ForwardListing,
+//!     request: ForwardListing,
 //! ) -> DeltaResult<Cursor<ForwardListing>> {
 //!     // Connector creates its cursor state and may begin prefetching here.
-//!     # Ok(Cursor::id(0))
+//!     todo!()
 //! }
 //!
 //! async fn advance_listing(
-//!     _cursor: Cursor<ForwardListing>,
+//!     cursor: Cursor<ForwardListing>,
 //! ) -> DeltaResult<Page<ForwardListing>> {
 //!     // Connector resolves its cursor state and fetches the next page here.
-//!     # Ok(Page {
-//!     #     data: vec![],
-//!     #     next: None,
-//!     # })
+//!     todo!()
 //! }
 //!
-//! // Connector decides how to dispatch kernel requests
+//! // Connector decides how to dispatch kernel requests. Here it uses async scheduling.
 //! async fn serve_request<N: Send + 'static>(
 //!     request: Request<N>,
 //! ) -> DeltaResult<N> {
@@ -90,7 +197,8 @@
 //!     }
 //! }
 //!
-//! // Connector also decides how to drive workflows.
+//! // Connector also decides how to drive workflows. Here, a generic driver loop. It could also
+//! // inject cancellation checks, logging, etc. if desired.
 //! async fn connector_workflow_driver<T: Send + 'static>(
 //!     mut workflow: DeltaResult<Workflow<T>>,
 //! ) -> DeltaResult<T> {
@@ -102,18 +210,11 @@
 //!     }
 //! }
 //!
+//! // A convenient connector-side entry point for the workflow, that hides all of the above.
 //! async fn connector_entry_point() -> DeltaResult<Foo> {
 //!     connector_workflow_driver(public_kernel_entry_point()).await
 //! }
 //! ```
-//!
-//! Kernel coroutines are an async-friendly synchronous API. Between suspensions, kernel performs
-//! only light, bounded CPU work. It must delegate all blocking, heavy, or unbounded work to the
-//! connector as a [`Request`]; failure to do so is a kernel bug. Connectors therefore do not need
-//! to treat kernel calls as scheduling points. Resume handles are [`Send`] and may move freely
-//! between tasks and threads as the workflow progresses. The connector may execute requests
-//! synchronously, asynchronously, on threads, or across distributed workers using the runtimes and
-//! libraries of its choosing.
 //!
 //! # Pagination
 //!
@@ -153,13 +254,13 @@
 //!
 //! - Starting a coroutine returns `DeltaResult<Workflow<_>>` or `DeltaResult<Generator<_, _>>`. An
 //!   error means the workflow failed before producing the next state.
-//! - [`TypedResume::resume`] accepts `DeltaResult<R>`. A connector error is delivered to the
+//! - The [`Resume`] function accepts `DeltaResult<R>`. A connector error is delivered to the
 //!   suspended kernel await point, where kernel may handle it or propagate it.
-//! - `resume` itself returns `DeltaResult`. An error means the workflow failed while kernel was
+//! - [`Resume`] itself returns `DeltaResult`. An error means the workflow failed while kernel was
 //!   processing the response or advancing the coroutine.
 //!
-//! A connector may drop any resume handle at any time. This abandons the suspended coroutine and
-//! drops all associated state. For connector-side work failures, resume with `Err` when kernel
+//! A connector may drop any [`Resume`] handle at any time. This abandons the suspended coroutine
+//! and drops all associated state. For connector-side work failures, resume with `Err` when kernel
 //! should observe the failure; otherwise drop the handle and return the error directly.
 // === Kernel implementation ===
 //
@@ -251,29 +352,25 @@
 
 use std::any::Any;
 use std::marker::PhantomData;
-use std::ops::Deref;
-use std::sync::{Arc, Weak};
 
 use bytes::Bytes;
-use tracing::{error, Instrument as _, Span};
-#[cfg(test)]
-use url::Url;
+use delta_kernel_derive::internal_api;
 
-pub(crate) use self::core::DeltaFuture;
-use self::core::{
-    advance_generator, advance_workflow, Exchange, Mailbox, PendingRequest, Wait, YieldExchange,
-};
+#[internal_api]
+pub(crate) use self::kernel::{Channel, DeltaFuture, GeneratorState, Yielder};
 use self::listing::{BackwardListing, BackwardListingResult, ForwardListing, ListingBounds};
 use self::read::{ReadJsonFiles, ReadParquetFiles};
 use self::write::WriteBytes;
 #[cfg(feature = "declarative-plans")]
 pub(crate) use crate::plans::Operation as PlanOperation;
-#[cfg(test)]
-use crate::Error;
 use crate::{DeltaResult, FileMeta, FileSlice, ParquetFooter};
 
 mod core;
 pub(crate) mod engine;
+#[cfg(feature = "internal-api")]
+pub mod kernel;
+#[cfg(not(feature = "internal-api"))]
+pub(crate) mod kernel;
 pub mod listing;
 pub mod read;
 pub mod write;
@@ -344,52 +441,11 @@ pub enum Workflow<O: Send + 'static> {
     Request(Request<Self>),
 }
 
-impl<O: Send + 'static> Workflow<O> {
-    /// Start a workflow and run it until completion or its first connector request.
-    pub(crate) fn start<Fut>(workflow: impl FnOnce(Channel) -> Fut) -> DeltaResult<Self>
-    where
-        Fut: DeltaFuture<O> + 'static,
-    {
-        let mailbox = Arc::new(Mailbox::default());
-        let future = workflow(Channel(Arc::clone(&mailbox)));
-        let future = async move {
-            future
-                .await
-                .inspect_err(|err| error!(error = %err, "kernel workflow failed"))
-        };
-        let task = Box::pin(future.instrument(Span::current()));
-        advance_workflow(task, mailbox)
-    }
-}
-
 /// A connector-facing generator that can complete, yield an item, or request connector work.
 pub enum Generator<O: Send + 'static, Y: Send + 'static> {
     Done(O),
     Yield(Y, YieldResume<Self>),
     Request(Request<Self>),
-}
-
-impl<O: Send + 'static, Y: Send + 'static> Generator<O, Y> {
-    /// Start a generator and run it until completion, its first yield, or a connector request.
-    pub(crate) fn start<Fut>(generator: impl FnOnce(Yielder<Y>) -> Fut) -> DeltaResult<Self>
-    where
-        Fut: DeltaFuture<O> + 'static,
-    {
-        let mailbox = Arc::new(Mailbox::default());
-        let yields = Arc::new(Mailbox::default());
-        let yielder = Yielder {
-            channel: Channel(Arc::clone(&mailbox)),
-            mailbox: Arc::clone(&yields),
-        };
-        let future = generator(yielder);
-        let future = async move {
-            future
-                .await
-                .inspect_err(|err| error!(error = %err, "kernel generator failed"))
-        };
-        let task = Box::pin(future.instrument(Span::current()));
-        advance_generator(task, mailbox, yields)
-    }
 }
 
 /// Operations kernel can request.
@@ -453,81 +509,3 @@ impl<N: Send + 'static> Request<N> {
 /// Resumes a suspended [`Workflow`] or [`Generator`] with connector's `response`, advancing it to
 /// the next communication point (suspend, yield, or completion).
 pub type Resume<N, R> = Box<dyn FnOnce(DeltaResult<R>) -> DeltaResult<N> + Send>;
-
-/// Kernel-side handle for typed connector operations.
-///
-/// It shares a mailbox with the coroutine driver and admits one live request at a time.
-pub(crate) struct Channel(Arc<Mailbox<PendingRequest>>);
-
-impl Channel {
-    /// Initiate a request/response exchange with the connector
-    async fn exchange<Out: Send + 'static, In: Send + 'static>(
-        &self,
-        outbound: Out,
-        pending: impl FnOnce(Weak<Exchange<Out, In>>) -> PendingRequest + Send,
-    ) -> DeltaResult<In> {
-        let exchange = Arc::new(Exchange::new(outbound));
-        self.0.publish(pending(Arc::downgrade(&exchange)))?;
-        Wait(exchange).await
-    }
-}
-
-/// Kernel-side handle passed to a generator body.
-///
-/// Forward all requests to the underlying [`Channel`] so connector can handle them, while
-/// delivering the input of [`Self::yield_item`] calls to the generator's immediate consumer.
-pub(crate) struct Yielder<Y> {
-    channel: Channel,
-    mailbox: Arc<Mailbox<Weak<YieldExchange<Y>>>>,
-}
-
-impl<Y: Send + 'static> Yielder<Y> {
-    /// Yield one item and suspend until the consumer resumes the generator.
-    ///
-    /// An error supplied by the consumer is returned at this await point.
-    pub(crate) async fn yield_item(&self, item: Y) -> DeltaResult<()> {
-        let emission = Arc::new(Exchange::new(item));
-        self.mailbox.publish(Arc::downgrade(&emission))?;
-        Wait(emission).await
-    }
-}
-
-impl<Y> Deref for Yielder<Y> {
-    type Target = Channel;
-
-    fn deref(&self) -> &Self::Target {
-        &self.channel
-    }
-}
-
-/// Kernel-side state for consuming a child generator.
-pub(crate) enum GeneratorState<W> {
-    Start(W),
-    Continue(YieldResume<W>),
-    Exhausted,
-}
-
-impl<Y: Send + 'static> GeneratorState<Generator<(), Y>> {
-    /// Return the next yielded item, forwarding connector requests through `parent`.
-    ///
-    /// Returns `None` after the child generator completes.
-    pub(crate) async fn next(&mut self, parent: &Channel) -> DeltaResult<Option<Y>> {
-        let state = std::mem::replace(self, Self::Exhausted);
-        let mut generator = match state {
-            Self::Start(generator) => Ok(generator),
-            Self::Continue(resume) => resume(Ok(())),
-            Self::Exhausted => return Ok(None),
-        };
-
-        loop {
-            generator = match generator? {
-                Generator::Request(request) => request.forward_to(parent).await,
-                Generator::Done(()) => return Ok(None),
-                Generator::Yield(item, resume) => {
-                    *self = Self::Continue(resume);
-                    return Ok(Some(item));
-                }
-            };
-        }
-    }
-}
